requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
reportlab==4.0.7
Flask-Mail==0.10.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
pytest==7.4.3
email-validator==2.1.0.post1
